+++ CurrencySpender/Helpers/VersionHelper.cs
@@ -42,6 +42,12 @@ namespace CurrencySpender.Helpers
                 C.SelectedCurrencies.Add(45690);
                 C.SelectedCurrencies.Add(48146);
             }
+            if (LowerVersionThan("1.2.4"))
+            {
+                C.SelectedCurrencies.Add(45690);
+                C.SelectedCurrencies.Add(45691);
+                C.SelectedCurrencies.Add(48146);
+            }
             if (C.Version == "0.0.0") return;
             if (LowerVersionThan(GetVersion()))
             {